将数据层使用的时间单位修改为毫秒 (#24)

+++ Cyan-Stars/Assets/GamePlay/Scritps/Data/MusicTimeline/ClipData.cs
@@ -9,9 +9,9 @@ using UnityEngine;
 public class ClipData
 {
     /// <summary>
-    /// 开始时间
+    /// 开始时间（毫秒）
     /// </summary>
-    public float StartTime;
+    public int StartTime;
     
     /// <summary>
     /// 速率
